+++ src/components/features/daos/view/tabs/gates-tab/gates-tab.tsx
