+++ src/components/features/daos/view/tabs/gates-tab/index.ts
