+++ src/components/features/daos/view/tabs/gates-tab/table-view.tsx
