+++ src/components/features/daos/view/tabs/static-grid-view-tab.tsx
